pyproject.toml
[project]
name = "kakao-novel-generator"
version = "0.1.0"
description = "Automated Korean web novel generator with ouroboros-ai integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "litellm>=1.80.0",
    "httpx>=0.27.0",
]

# ouroboros-ai integration requires Python 3.14+
# Install separately with: pip install ouroboros-ai
# [project.optional-dependencies]
# ouroboros = ["ouroboros-ai>=0.5.0"]

[project.scripts]
novel = "novel_generator.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/novel_generator"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
strict = true
